# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ResidentVoice Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /buildings/slug/{slug}:
    get:
      operationId: getBuildingBySlug
      tags: [buildings]
      summary: Get building info by slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Building info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
        "404":
          description: Not found
  /buildings/{buildingId}:
    get:
      operationId: getBuilding
      tags: [buildings]
      summary: Get building details
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Building details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
    patch:
      operationId: updateBuilding
      tags: [buildings]
      summary: Update building info (admin)
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildingUpdate"
      responses:
        "200":
          description: Updated building
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
  /buildings/{buildingId}/settings:
    get:
      operationId: getBuildingSettings
      tags: [buildings]
      summary: Get building settings
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Building settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildingSettings"
    patch:
      operationId: updateBuildingSettings
      tags: [buildings]
      summary: Update building settings (admin)
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildingSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildingSettings"

  # ─── Residents ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Building:
      type: object
      required: [id, name, slug, city, country, status, plan, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: ["string", "null"]
        country:
          type: string
        timezone:
          type: ["string", "null"]
        numberOfUnits:
          type: ["integer", "null"]
        numberOfFloors:
          type: ["integer", "null"]
        buildingType:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, pending, suspended]
        plan:
          type: string
          enum: [free, basic, pro, enterprise]
        createdAt:
          type: string
    BuildingUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        timezone:
          type: string
        numberOfUnits:
          type: integer
        numberOfFloors:
          type: integer
        buildingType:
          type: string
        logoUrl:
          type: string
        photoUrl:
          type: string
    BuildingSettings:
      type: object
      required: [id, buildingId]
      properties:
        id:
          type: integer
        buildingId:
          type: integer
        verificationMethod:
          type: string
          enum: [manual_approval, invite_code, document_upload]
        allowTenantOfficialVotes:
          type: boolean
        oneVotePerUnit:
          type: boolean
        passPercentage:
          type: integer
        allowAnonymousPosts:
          type: boolean
        inviteCode:
          type: ["string", "null"]
    BuildingSettingsUpdate:
      type: object
      properties:
        verificationMethod:
          type: string
        allowTenantOfficialVotes:
          type: boolean
        oneVotePerUnit:
          type: boolean
        passPercentage:
          type: integer
        allowAnonymousPosts:
          type: boolean
        inviteCode:
          type: string
...